requirements.txt
fastapi 
uvicorn 
python-multipart 
torchaudio 
pyannote.audio 
scipy
